Fixed metric object hierarchy
Also some test files fixes

+++ Rocnikovy_projekt/Code/Progect/metric_value.py
@@ -1,6 +1,6 @@
 from datetime import datetime
 class MetricValue:
-    def __init__(self, metric_name: str, value: int, timestamp: datetime):
+    def __init__(self, metric_name: str, value: float, timestamp: datetime):
         self.metric_name = metric_name
         self.value = value
         self.timestamp = timestamp
